Dynamically Generated Path-Finding
In previous commits, I was enabling a ground tile in the navigation layer of the tileset. That was enabling navigation for thousands of tiles of that type, which is why I couldn't have more than one world's ground tile enabled, because it ended up adding too many tiles at once for navigation.

This new idea builds off the tile processing system I already have to dynamically enable tiles for path-finding as the player or characters encounter them. It works pretty well, and is transferrable across all worlds. It uses alternative tiles and tracks the environment beneath it, and it could be configured for the terrain but I might not. It will need to be configured for buildings too.

The next thing I could try to do for path-finding is to implement a limit to the number of path-finding tiles that can be enabled at once, and begin disabling the oldest ones when the limit is reached. After that, I can try to enable the food buddy's jumping abilities.