Performance Improvement: Tile Objects
There was an issue with Tile Objects being instantiated and loaded rather than just an object being created, so it was impacting performance whenever a character would stand near a bush or a building entrance for too long. I fixed this to be more efficient and to no longer instantiate an object unless one is being loaded into the game.

Now that so many of our assets are implemented, the game is getting larger and it is taking more processing time. There are several TileMapLayers in our game, and although some may start off disabled, many load in as the player explores the map, and the processing time skyrockets as this occurs. I need to now start focusing on the asset management I had thought out at the very beginning of this project.