Character Status Bars Implemented Graphically AND Functionally
The status bars for both the player and the two currently active food buddies are implemented properly with Godot UI, Ryan's graphics, and my code functionality. The stamina and XP are confirmed to be properly tracked, and the health almost certainly is too, but I didn't get to check because the enemy can't attack currently (needs to have path-finding enabled still).