More Bug Fixes: Jumping
I fixed the bugs related to jumping near environment assets such that they can't be jumped through anymore but can still be jumped near. There is now a jump_enabled bool that is set to false (disabled) when the character is touching an environmentasset's collider. It is only re-enabled when the character is no longer touching it. Also, if a character runs and jumps towards a tree, the jump will be cancelled if the character is trying to move down towards the tree (so that the jump not ending while moving downward glitch could be fixed).