Field State Interface UI & Functionality: COMPLETE
I implemented the UI for the food buddy field state interface in Godot, and I didn't end up using most of my existing functionality for it. I removed the ability to control it with WASD because that wouldn't be consistent with the other interfaces. You click on the field state that you want the food buddy to have selected. I did keep all of the triggering logic for actually toggling the interface, and I replicated that for the food buddy select yesterday too actually.

The only thing this needs now is a hover skin from Ryan.