Dialogue System Dynamic Play Conditions
I implemented the dynamic play conditions into the dialogue system. I started a 'play_parser' dictionary that holds string keys like "player_level", and the value is a corresponding callback function that will execute taking in the play condition string as an argument to parse. It will evaluate the condition and return true or false.

I have a 'play_when' condition as well, which can be either 'now' or 'user'. This condition can only be reached if all play conditions were met. If 'now', the conversation will be played immediately- even if all conversations haven't been checked yet. If 'user', the conversation will be saved to a list for the user to select from if no conversations auto-play- however, if a conversation hasn't played yet, it will auto-play the first time its considered despite having the 'user' tag.