Dialogue Interface & Selection UI and Functionality COMPLETE
I added the ability to select from a list of conversation options. I refined all of the UI to work smoothly and function effectively. PLAY and PLAY_WHEN conditions are functionally implemented and new parse conditions can easily be implemented. DIALOGUE and GAME instructions could be implemented pretty easily, with some minor work and parse instruction creations. But even so, I'm not going to implement those because I don't think it's necessary for the capstone demo. That can be a long term implementation