Juicebox Functionality: COMPLETE
Juiceboxes are fully animated and fully functional. They throw to wherever the player clicks as long as they have juiceboxes to throw and as long as they are in juice mode by pressing 'C'.

The juiceboxes heal any food buddies or the player that are in range of the explosion. It is friggin dope.