Food Buddy Foraging: FUNCTIONAL
I got food buddy foraging to work, so they are able to forage independently. There are some things that need improvement.

The way that path-finding works currently is that paths are built in places you move, so the world starts empty with no paths. This is causing some problems with the food buddies ability to manuever around bushes- sometimes they'll get stuck underneath them because they don't see the path to walk around them.

There aren't any animations for the foraging, it'd be cool if they could rustle the bush. It'd also be useful for each food buddy to have a berry, juice, and juicebox counter next to their other status details.

There is no cooldown timer when they drop off berries to Brittany, there should be about a second where they idle as if they are taking the time to deposit the berries.

also updated "solo" to "fight" fieldstate