Sausage Link Code Preparation for Animations
I made a sort of state machine that I think will work with sausage link's animations, depending on proper naming schemes. It automatically and dynamically determines the direction the animation should face using a dictionary of vector2 keys with the current x and y directions, with the values being constants of "back", "front", or "sideways".
The states, "idle", "moving", "ability", and "die", are set through specific if-statements.