Food Buddy World Navigation Improvements
Since I won't be able to get jumping onto terrains working in time for the game's release, I decided to make sure that the food buddies can at least avoid the terrain properly altogether.

When they are in the follow or fight field state and are following the player, they will stop moving if the player's altitude increases.

Additionally, I increased the width and height of the terrain tiles to 2 in the tilemap, so there is basically a layer of padding between them and the food buddies at all times. This ensures that the food buddies can turn corners around terrain without colliding and glitching.

I commented out and disabled any code I had for food buddy jumping, as it wasn't quite working where I left it off. Even if I could get it working by the deadline, I don't think it'd be very refined and I'd lose time in other areas.