Dan the Donut & Animation State Machine COMPLETE
Dan the Donut's animations, movements, and ability are fully functional and they look amazing. We are very happy with them.

In this commit I also fixed the animation state machine to properly determine the directions the animations should play in- as I FINALLY discovered how to write a lambda in GDScript (absolute gamechanger, cant WAIT to redo this codebase)