Interface Bug Fixing
I improved the interface logic such that all assets and characters in the game are effectively paused while the interface is open. If food buddies are moved to the center of the interface (select and field state), they are transported back to their previous locations immediately when the interface closes.

I'm currently working on the berry bot interface's animations so that the berries smoothly transition to their progress points.