Quick Interface Bug Fixes (hopefully the last of em)
In the berry bot interface, if a character was dead, their death animation would replay upon closing the interface. I made sure it only replays if they're alive.

I also added different variants of the food buddy select icon, so it is finally clear as to which ones are selected while using that interface.