Updated The Garden and The Center
I updated The Garden and The Center. I've removed all climbable terrains, as the jumping and elevation didn't end up being a feature that was utilized- and it was never a requirement I'd listed for this demo in the first place. I made several winding and connecting pathways with little forests of trees, bushes, and rocks in between- each spaced far enough apart for adequate path-finding.

Batter Badlands is up next, then the whole map will be ready and bug-free. Next up I'll have to connect the music to each of the worlds and trigger it to change upon switching worlds, and then I'll need to implement more buildings into each village, add some NPCs into the houses and town areas, trigger enemy spawns, then the map is complete!