Brittany Fusion Functionality
I implemented the functionality to get Player-FoodBuddy fusions working, and it is currently working well with Brittany. When a food buddy is equipped, the "Fuse!" animation is played, and as it finishes the food buddy is disabled and hidden and the player's animation is updated to have the food buddy. The moveset is based on the food buddy's abiltiies that can be written.