Buildings Work
I made some updates to buildings so that they have a custom data for the offset that the player should be adjusted when entering/exiting a building. This is a dynamic fix.

I need to figure out how to enable collisions for food buddies and enemies with buildings. For some reason even though it seems the collison values are being set, they aren't colliding.