Batter Badlands Map Environment Finished & Building Entering/Exiting Perfected
I finished up the batter badlands environment with a cool rock structure towards the bottom right of the map. I still need to add buildings, as I do for all of the other worlds, so each world is up to the same development state now.

I also implemented the proper tilemap logic and code to allow food buddies, food enemies, and food citizens to properly navigate inside and outside of buildings, so the transition between entering and exiting is much smoother now.

Collisions are now also finally working with buildings.