HUGE PERFORMANCE BOOST
A combination of things that I've been trying to do for a while have finally come together nicely. I made it so that the path-finding tiles are dynamically removed once they reach a cap of 500 tiles at once. This ensures that there is enough space of characters to freely move while they generate new tiles and paths, but it prevents the navigation space from growing too vast for quick calculations.

This moved the game from 60 fps while having 3 enemies and 2 food buddies loaded in to 120 fps while having 9 enemies in. The big change that made the fps upgrade possible came after the path-finding upgrade and the removal of jumping and terrains from the demo. Those factors allowed me to be able to delay the process_nearby_tiles function calls for all of the characters, so it didn't have to run every frame anymore since the characters aren't needing that precise of updates anymore.