MAP VISUALS ALL COMPLETE
I've implemented all of the necessary tile structure, data, and placements in the map- so it is now complete. Every world has houses of two different sizes placed, and they each function identically. I still need to add characters into some of the houses, but I plan to do that a little later on (along with the dialogue for characters).