Forage & More General Map Bug Fixing
I fixed foraging so that food buddies share knowledge of which bushes they've already foraged, and they're smart enough to not go after the same bush. They choose bushes closest to the player. After foraging a bush, they won't revisit it until they forage another 20 other bushes.

The bugs are almost all gone right now, and the only ones I know of that remain are a rare one where Dan doesn't unfuse properly- he stays in the PLAYER state and then stays broken. Also, food buddies are not cooperating with the houses very nicely. I either need to scale back on the houses and only do them towards the tops of the map, or I need to find a solution for this before demo due date.