Player-Link Animations In & Player Look At Mouse
When the player attacks, the character will look in the direction that the mouse is pointed. Link's attacks are working, but I need to switch the hitbox that is being used. This might require me to create separate "health" hitboxes which will be the ones that actually receive damage, vs "damage" hitboxes which disperse damage.

There's a few bugs right now- sometimes when switching in and out of the PLAYER state, the food buddy glitches and stays stuck in the PLAYER state. Additionally, if a fusion occurs during a level-up or another event, it can cause glitches. I should make sure that the fusion process is paused when interfaces open.
Path-finding is working well now and I understand how expensive it is. More importantly, I know how to fine tune it, so I did so to the best of my abilities. In order for the characters to have a full range of the right paths, they need to process more tiles around them at once. Maybe I can do it in a different, less expensive way that lets them process the same or even larger/wider space. Gonna try that now