Tile Processing Updates IMPORTANT NOTES FOR FUTURE
I updated the tile manager such that it only creates actual tile instances when updating the navigation or calling a tile callback function. I think I could modify the path-finding logic eventually into a system that find tiles that CAN'T be traveled on and build a perimeter of tiles around them so that characters know how to maneuver around them.