Path-Finding Tile Processing & Enemy Processing
I made another (hopefully final for this demo) improvement to path-finding.; The characters now generate rings of path-finding tiles, and these rings are long and tall enough to get around most objects. The only thing they still get stuck on are fences and houses, which I want to fix, but I don't know if its best to really prioritize that.

Enemies now only process while on screen, and their visual on screen notifiers constantly update to dynamically enable/disable the enemy's processing when they enter/exit the screen. I can also modify this to only disable processing if they aren't actually engaged with a food buddy or if they're hot on the player's tail.