Title Screen to Game Transition & Animation Spritesheets
I improved the title screen and working on a smooth transition into the game. I also had to fix up the music transitions because of the world-altering stuff I did last commit.

I made the spritesheets for the enemy animations that Ryan drew today and the frames for Sir Loin.